+++ src/app/Components/remove-group-from-course/remove-group-from-course.component.ts
@@ -1,4 +1,4 @@
-import { Component, OnInit } from '@angular/core';
+import {Component, OnInit} from '@angular/core';
 import {ParticipantsService} from "../../Services/ParticipantsService/participants.service";
 import {NotificationService} from "../../Services/NotificationService/notification.service";
 import {FormBuilder, Validators} from "@angular/forms";
@@ -10,12 +10,13 @@ import {FormBuilder, Validators} from "@angular/forms";
 })
 export class RemoveGroupFromCourseComponent implements OnInit {
 
-  constructor(private readonly participantsService: ParticipantsService,private notifyService: NotificationService,private fb: FormBuilder) { }
+  constructor(private readonly participantsService: ParticipantsService, private notifyService: NotificationService, private fb: FormBuilder) {
+  }
 
   removeGroupOfStudentsFromCourseForm = this.fb.group({
 
-    groupNumber: ['', Validators.required],
-    courseId: ['', Validators.required]
+    groupNumber: [null, Validators.required],
+    courseId: [null, Validators.required]
   })
 
   ngOnInit(): void {
@@ -24,11 +25,17 @@ export class RemoveGroupFromCourseComponent implements OnInit {
   removeGroupOfStudentsFromCourse(): void {
     let groupNumber = this.removeGroupOfStudentsFromCourseForm.value.groupNumber;
     let courseId = this.removeGroupOfStudentsFromCourseForm.value.courseId;
-    this.removeGroupOfStudentsFromCourseForm.reset();
+
+    if (groupNumber == null || courseId == null)
+      throw this.notifyService.showError("All fields are required");
+
     this.participantsService.removeGroupFromCourse(groupNumber, courseId).subscribe((data: any) => {
-    }, (err) => {
-      this.notifyService.showError(err.error.message);
-    });
+      }, (err) => {
+        this.notifyService.showError(err.error.message);
+      },
+      () => {
+        this.notifyService.showSuccess("Success");
+      });
   }
 
 }
